Simplify the base serializer
