+++ src/django/api/serializers/facility/facility_download_serializer_base.py
@@ -1,37 +1,17 @@
-import copy
-from typing import Any, Dict, List
+from typing import Any, List
 from api.models.facility.facility_index import FacilityIndex
 from api.models.facility.facility_manager_index_new import (
     FacilityIndexNewManager,
 )
 from countries.lib.countries import COUNTRY_NAMES
-from api.csv_download import (
-    format_download_extended_fields,
-    format_download_claimed_fields,
-    CLAIMED_DOWNLOAD_FIELDS,
-)
-from api.constants import CLAIMED_DOWNLOAD_FIELDS_MAPPING
+from api.csv_download import format_download_extended_fields
 from api.helpers.helpers import parse_download_date
-from api.serializers.facility.mit_living_wage_download_helper import (
-    MIT_LIVING_WAGE_DOWNLOAD_HEADERS,
-    MITLivingWageDownloadHelper,
-)
-from api.serializers.facility.partner_field_helper import (
-    build_object_field_cells,
-    build_primitive_field_cell,
-    get_cached_all_partner_fields,
-    group_extended_fields_by_name,
-    partner_field_property_paths,
-    apply_schema_defaults,
-)
-from api.serializers.facility.wage_indicator_download_helper import (
-    WAGE_INDICATOR_DOWNLOAD_HEADERS,
-    WageIndicatorDownloadHelper,
-)
 from rest_framework.serializers import Serializer, SerializerMethodField
 
 
 class FacilityDownloadSerializerBase(Serializer):
+    """Shared CSV row helpers for full and embed facility downloads."""
+
     row = SerializerMethodField()
 
     class Meta:
@@ -59,32 +39,8 @@ class FacilityDownloadSerializerBase(Serializer):
     ]
     IS_CLOSED_HEADER = "is_closed"
 
-    CLAIMED_FIELDS_HEADERS = [header for header, _ in CLAIMED_DOWNLOAD_FIELDS_MAPPING]
-
-    PARTNER_FIELD_HEADER_SEPARATOR = "."
-    PARTNER_FIELD_VALUE_SEPARATOR = "|"
-
-    def __init__(self, *args, **kwargs) -> None:
-        super().__init__(*args, **kwargs)
-        # Built once per serializer instance so each system-partner-field
-        # provider lookup is reused across every row in a paginated
-        # download instead of being rebuilt per facility.
-        self.__mit_living_wage_helper = MITLivingWageDownloadHelper()
-        self.__wage_indicator_helper = WageIndicatorDownloadHelper()
-        self.__partner_fields = [
-            field
-            for field in get_cached_all_partner_fields()
-            if field.active and field.available_in_downloads
-        ]
-        self.__partner_fields.sort(key=lambda field: field.name)
-        self.__default_partner_fields = [
-            field for field in self.__partner_fields if not field.system_field
-        ]
-        self.__system_partner_fields = {
-            field.name for field in self.__partner_fields if field.system_field
-        }
-
     def get_common_row(self, facility: FacilityIndexNewManager):
+        """Return the leading columns shared by every download format."""
         return [
             facility.id,
             self.get_contribution_date(facility),
@@ -98,11 +54,13 @@ class FacilityDownloadSerializerBase(Serializer):
 
     @staticmethod
     def get_contribution_date(facility: FacilityIndexNewManager) -> str:
+        """Return the facility creation date formatted for CSV export."""
         created_at = facility.created_from_info["created_at"]
         return parse_download_date(created_at)
 
     @staticmethod
     def get_name(facility: FacilityIndexNewManager) -> str:
+        """Return the English claim name when present, otherwise the facility name."""
         claim = facility.approved_claim
         if claim and claim.get("facility_name_english"):
             return claim.get("facility_name_english")
@@ -110,164 +68,31 @@ class FacilityDownloadSerializerBase(Serializer):
 
     @staticmethod
     def get_country_name(facility: FacilityIndexNewManager) -> str:
+        """Return the display name for the facility country code."""
         return COUNTRY_NAMES.get(facility.country_code, "")
 
     @staticmethod
     def get_location(facility: FacilityIndexNewManager) -> List[str]:
+        """Return latitude and longitude as ``[lat, lng]``."""
         return list(reversed(facility.location.coords))
 
     def get_sector(self, facility: FacilityIndex):
+        """Return pipe-separated, title-cased sector values."""
         return self.join(facility.sector)
 
     def get_extended_fields(self, fields) -> List[Any]:
+        """Return the six standard extended-field columns, joined with ``|``."""
         extended_fields = [[], [], [], [], [], []]
         format_download_extended_fields(fields, extended_fields)
 
         return list(map("|".join, extended_fields))
 
-    def get_claimed_fields(self, facility: FacilityIndexNewManager) -> List[str]:
-        claimed_fields = [""] * len(CLAIMED_DOWNLOAD_FIELDS)
-        claim = facility.approved_claim
-        if claim:
-            format_download_claimed_fields(claim, claimed_fields)
-        return claimed_fields
-
     @staticmethod
     def get_is_closed(facility: FacilityIndexNewManager) -> str:
+        """Return ``"True"`` or ``"False"`` for the facility closed status."""
         return str(facility.is_closed if facility.is_closed is not None else False)
 
     @staticmethod
     def join(arr: List[Any]) -> str:
+        """Join list items with ``|``, applying title case to each."""
         return "|".join([item.title() for item in arr])
-
-    def get_partner_fields_headers(self) -> List[str]:
-        """
-        Build CSV headers for the provided active partner fields.
-
-        - Object-typed partner fields with a JSON Schema expand into one
-          column per leaf property, joining the path with ".". Nested
-          objects are walked recursively until a non-object leaf is
-          reached (e.g. "amfori.bsci_audit.submission_date").
-        - Any other partner field (or missing schema) emits a single
-          column named after the partner field itself.
-        """
-        headers: List[str] = []
-        for partner_field in self.__default_partner_fields:
-            paths = partner_field_property_paths(partner_field)
-            if paths:
-                for path in paths:
-                    headers.append(
-                        self.PARTNER_FIELD_HEADER_SEPARATOR.join(
-                            (partner_field.name, *path)
-                        )
-                    )
-            else:
-                headers.append(partner_field.name)
-        return headers
-
-    def get_partner_fields_row(
-        self,
-        extended_fields: List[Dict[str, Any]],
-    ) -> List[str]:
-        """
-        Build CSV row cells for the provided active partner fields by
-        matching each partner field's `name` against entries in the
-        facility's `extended_fields`.
-
-        For object-typed partner fields with a JSON Schema each stored
-        `raw_values` dict is walked following the schema leaf paths so
-        the output columns stay aligned with `get_partner_fields_headers`.
-        Multiple contributions for the same partner field on one facility
-        are joined with "|" per leaf column (consistent with the existing
-        extended-field behavior).
-        """
-        grouped = group_extended_fields_by_name(extended_fields)
-        separator = self.PARTNER_FIELD_VALUE_SEPARATOR
-        row: List[str] = []
-
-        for partner_field in self.__default_partner_fields:
-            entries = copy.deepcopy(grouped.get(partner_field.name, []))
-
-            for entry in entries:
-                if not isinstance(entry, dict):
-                    continue
-
-                value = entry.get("value", {})
-
-                if not isinstance(value, dict):
-                    continue
-
-                if "raw_values" not in value:
-                    continue
-
-                raw_values = value.get("raw_values", {})
-
-                if not isinstance(raw_values, dict):
-                    continue
-
-                schema = partner_field.json_schema
-
-                if not isinstance(schema, dict):
-                    continue
-
-                entry["value"]["raw_values"] = apply_schema_defaults(
-                    raw_values,
-                    schema,
-                )
-
-            paths = partner_field_property_paths(partner_field)
-
-            if paths:
-                row.extend(build_object_field_cells(entries, paths, separator))
-            else:
-                row.append(build_primitive_field_cell(entries, separator))
-
-        return row
-
-    def get_mit_living_wage_headers(self) -> List[str]:
-        """
-        Download-only headers for the `mit_living_wage` system partner
-        field. Kept separate from `get_partner_fields_headers` because
-        `mit_living_wage` is filtered out of the regular active-partner
-        pipeline (it has `system_field=True`) and the two columns are
-        synthesized at download time from the provider + partner-field
-        row, not from `facility.extended_fields`.
-        """
-        if "mit_living_wage" not in self.__system_partner_fields:
-            return []
-        return MIT_LIVING_WAGE_DOWNLOAD_HEADERS
-
-    def get_mit_living_wage_row(self, facility: FacilityIndexNewManager) -> List[str]:
-        """
-        Download-only cells matching `get_mit_living_wage_headers`.
-        Returns two empty strings for any facility the MIT provider
-        can't resolve (non-US/PR/VI, no TIGER/Line match, missing
-        partner-field row, etc.).
-        """
-        if "mit_living_wage" not in self.__system_partner_fields:
-            return []
-        return self.__mit_living_wage_helper.get_cells(facility)
-
-    def get_wage_indicator_headers(self) -> List[str]:
-        """
-        Download-only headers for the `wage_indicator` system partner
-        field. Like `mit_living_wage`, `wage_indicator` has
-        `system_field=True` so it's excluded from the regular
-        active-partner pipeline; its six columns (URL + display text
-        for each of the three link types) are synthesized at download
-        time from `WageIndicatorProvider` output.
-        """
-        if "wage_indicator" not in self.__system_partner_fields:
-            return []
-        return WAGE_INDICATOR_DOWNLOAD_HEADERS
-
-    def get_wage_indicator_row(self, facility: FacilityIndexNewManager) -> List[str]:
-        """
-        Download-only cells matching `get_wage_indicator_headers`.
-        Returns six empty strings for any facility whose country has
-        no `WageIndicatorCountryData` row, or when the
-        `wage_indicator` `PartnerField` has no associated contributor.
-        """
-        if "wage_indicator" not in self.__system_partner_fields:
-            return []
-        return self.__wage_indicator_helper.get_cells(facility)
